+++ src/extension.ts
@@ -1,10 +1,12 @@
 import { Marp } from '@marp-team/marp-core'
-import { ExtensionContext, commands, workspace } from 'vscode'
+import { ExtensionContext, Uri, commands, workspace } from 'vscode'
 import exportCommand from './commands/export' // tslint:disable-line: import-name
 import showQuickPick from './commands/show-quick-pick'
+import customTheme from './plugins/custom-theme'
 import lineNumber from './plugins/line-number'
 import outline from './plugins/outline'
 import { marpCoreOptionForPreview, clearMarpCoreOptionCache } from './option'
+import themes from './themes'
 
 const frontMatterRegex = /^-{3,}\s*([^]*?)^\s*-{3}/m
 const marpDirectiveRegex = /^marp\s*:\s*true\s*$/m
@@ -12,6 +14,7 @@ const marpVscode = Symbol('marp-vscode')
 const shouldRefreshConfs = [
   'markdown.marp.breaks',
   'markdown.marp.enableHtml',
+  'markdown.marp.themes',
   'markdown.preview.breaks',
   'window.zoomLevel', // for WebKit polyfill
 ]
@@ -28,21 +31,50 @@ export function extendMarkdownIt(md: any) {
   md.parse = (markdown: string, env: any) => {
     // Generate tokens by Marp if enabled
     if (detectMarpFromFrontMatter(markdown)) {
-      md[marpVscode] = new Marp(marpCoreOptionForPreview(md.options))
+      const mdFolder = Uri.parse(md.normalizeLink('.')).with({ scheme: 'file' })
+      const workspaceFolder = workspace.getWorkspaceFolder(mdFolder)
+      const baseFolder = workspaceFolder ? workspaceFolder.uri : mdFolder
+
+      const marp = new Marp(marpCoreOptionForPreview(md.options))
+        .use(customTheme)
         .use(outline)
         .use(lineNumber)
 
-      const marpMarkdown = md[marpVscode].markdown
+      // Load custom themes
+      let shouldRefresh = false
+
+      Promise.all(
+        themes.loadStyles(baseFolder).map(promise =>
+          promise
+            .then(theme => {
+              if (theme.registered) shouldRefresh = true
+            })
+            .catch(e => console.error(e))
+        )
+      ).then(() => {
+        if (shouldRefresh) commands.executeCommand('markdown.preview.refresh')
+      })
+
+      for (const theme of themes.getRegisteredStyles(baseFolder)) {
+        try {
+          marp.themeSet.add(theme.css)
+        } catch (e) {
+          console.error(
+            `Failed to register custom theme from "${theme.path}". (${e.message})`
+          )
+        }
+      }
 
       // Use image stabilizer and link normalizer from VS Code
-      marpMarkdown.renderer.rules.image = md.renderer.rules.image
-      marpMarkdown.normalizeLink = md.normalizeLink
+      marp.markdown.renderer.rules.image = md.renderer.rules.image
+      marp.markdown.normalizeLink = md.normalizeLink
 
       // validateLink prefers Marp's default. If overridden by VS Code's it,
       // does not return compatible result with the other Marp tools.
-      // marpMarkdown.validateLink = md.validateLink
+      // marp.markdown.validateLink = md.validateLink
 
-      return marpMarkdown.parse(markdown, env)
+      md[marpVscode] = marp
+      return marp.markdown.parse(markdown, env)
     }
 
     // Fallback to original instance if Marp was not enabled
@@ -69,14 +101,9 @@ export function extendMarkdownIt(md: any) {
 
 export const activate = ({ subscriptions }: ExtensionContext) => {
   subscriptions.push(
-    commands.registerCommand('markdown.marp.export', exportCommand)
-  )
-
-  subscriptions.push(
-    commands.registerCommand('markdown.marp.showQuickPick', showQuickPick)
-  )
-
-  subscriptions.push(
+    commands.registerCommand('markdown.marp.export', exportCommand),
+    commands.registerCommand('markdown.marp.showQuickPick', showQuickPick),
+    themes,
     workspace.onDidChangeConfiguration(e => {
       if (shouldRefreshConfs.some(c => e.affectsConfiguration(c))) {
         clearMarpCoreOptionCache()